+++ src/main/java/com/jsharpe/plantlive/repositories/details/in/DetailInRepository.java
@@ -8,6 +8,6 @@ import java.util.Date;
 public interface DetailInRepository {
 
     @SqlUpdate("INSERT INTO details (plant_id, in_timestamp, temperature, humidity, light, conductivity) VALUES (?, ?, ?, ?, ?, ?)")
-    int save(long plantId, Date inTimestamp, int temperature, int humidity, int light, int conductivity) throws SQLException;
+    int save(long plantId, Date inTimestamp, double temperature, double humidity, double light, double conductivity) throws SQLException;
 
 }
